requirements.txt

boto3==1.34.11
Flask==2.3.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
psycopg2=2.9.9
python-dateutil==2.8.2
stripe==7.10.0